add more client functions #26

+++ src/powerkit_client.cpp
@@ -11,6 +11,7 @@
 
 double PowerClient::getBatteryLeft(QDBusInterface *iface)
 {
+    if (!iface) { return -1; }
     qDebug() << "check for battery left";
     if (!iface->isValid()) { return false; }
     QDBusMessage reply = iface->call("BatteryLeft");
@@ -21,6 +22,7 @@ double PowerClient::getBatteryLeft(QDBusInterface *iface)
 
 bool PowerClient::hasBattery(QDBusInterface *iface)
 {
+    if (!iface) { return false; }
     qDebug() << "check if we have any battery";
     if (!iface->isValid()) { return false; }
     QDBusMessage reply = iface->call("HasBattery");
@@ -31,6 +33,7 @@ bool PowerClient::hasBattery(QDBusInterface *iface)
 
 bool PowerClient::onBattery(QDBusInterface *iface)
 {
+    if (!iface) { return false; }
     qDebug() << "check if we are on battery";
     if (!iface->isValid()) { return false; }
     QDBusMessage reply = iface->call("OnBattery");
@@ -41,6 +44,7 @@ bool PowerClient::onBattery(QDBusInterface *iface)
 
 qlonglong PowerClient::timeToEmpty(QDBusInterface *iface)
 {
+    if (!iface) { return -1; }
     qDebug() << "check for time to empty";
     if (!iface->isValid()) { return false; }
     QDBusMessage reply = iface->call("TimeToEmpty");
@@ -51,6 +55,7 @@ qlonglong PowerClient::timeToEmpty(QDBusInterface *iface)
 
 bool PowerClient::canHibernate(QDBusInterface *iface)
 {
+    if (!iface) { return false; }
     qDebug() << "check if we can hibernate";
     if (!iface->isValid()) { return false; }
     QDBusMessage reply = iface->call("CanHibernate");
@@ -61,6 +66,7 @@ bool PowerClient::canHibernate(QDBusInterface *iface)
 
 bool PowerClient::canSuspend(QDBusInterface *iface)
 {
+    if (!iface) { return false; }
     qDebug() << "check if we can suspend";
     if (!iface->isValid()) { return false; }
     QDBusMessage reply = iface->call("CanSuspend");
@@ -71,6 +77,7 @@ bool PowerClient::canSuspend(QDBusInterface *iface)
 
 bool PowerClient::canRestart(QDBusInterface *iface)
 {
+    if (!iface) { return false; }
     qDebug() << "check if we can restart";
     if (!iface->isValid()) { return false; }
     QDBusMessage reply = iface->call("CanRestart");
@@ -81,6 +88,7 @@ bool PowerClient::canRestart(QDBusInterface *iface)
 
 bool PowerClient::canPowerOff(QDBusInterface *iface)
 {
+    if (!iface) { return false; }
     qDebug() << "check if we can poweroff";
     if (!iface->isValid()) { return false; }
     QDBusMessage reply = iface->call("CanPowerOff");
@@ -91,6 +99,7 @@ bool PowerClient::canPowerOff(QDBusInterface *iface)
 
 bool PowerClient::lidIsPresent(QDBusInterface *iface)
 {
+    if (!iface) { return false; }
     qDebug() << "check if we have a lid";
     if (!iface->isValid()) { return false; }
     QDBusMessage reply = iface->call("LidIsPresent");
@@ -101,9 +110,52 @@ bool PowerClient::lidIsPresent(QDBusInterface *iface)
 
 void PowerClient::lockScreen(QDBusInterface *iface)
 {
+    if (!iface) { return; }
     qDebug() << "lock screen";
     if (!iface->isValid()) { return; }
     QDBusMessage reply = iface->call("LockScreen");
     bool ok = reply.errorMessage().isEmpty();
     qDebug() << "locked screen?" << ok;
+}
+
+void PowerClient::hibernate(QDBusInterface *iface)
+{
+    if (!iface) { return; }
+    qDebug() << "hibernate";
+    if (!iface->isValid()) { return; }
+    QDBusMessage reply = iface->call("Hibernate");
+    bool ok = reply.errorMessage().isEmpty();
+    qDebug() << "reply" << ok;
+}
+
+void PowerClient::suspend(QDBusInterface *iface)
+{
+    if (!iface) { return; }
+    qDebug() << "suspend";
+    if (!iface->isValid()) { return; }
+    QDBusMessage reply = iface->call("Suspend");
+    bool ok = reply.errorMessage().isEmpty();
+    qDebug() << "reply" << ok;
+}
+
+bool PowerClient::restart(QDBusInterface *iface)
+{
+    if (!iface) { return false; }
+    qDebug() << "restart";
+    if (!iface->isValid()) { return false; }
+    QDBusMessage reply = iface->call("Restart");
+    bool ok = reply.errorMessage().isEmpty();
+    qDebug() << "reply" << ok;
+    return ok;
+}
+
+bool PowerClient::poweroff(QDBusInterface *iface)
+{
+    if (!iface) { return false; }
+    qDebug() << "poweroff";
+    if (!iface->isValid()) { return false; }
+    QDBusMessage reply = iface->call("PowerOff");
+    bool ok = reply.errorMessage().isEmpty();
+    qDebug() << "reply" << ok;
+    return ok;
 }
